p4i integration Integrate 182219 from 4.0 to main: Fixes:    Printing of images in richtext on an image paintdevice RevBy:    Brad Details:        When calculating the scaling factor from image resolution to paint device resolution        we need to use floating point division, not integer.
[git-p4: depot-paths = "//depot/qt/main/": change = 182220]

+++ src/gui/text/qtextimagehandler.cpp
@@ -60,9 +60,8 @@ static QPixmap getPixmap(QTextDocument *doc, const QTextImageFormat &format)
 
     qreal scale = 1.0;
     QPaintDevice *pdev = doc->documentLayout()->paintDevice();
-    if (pdev && !pm.isNull()) 
-        scale = pdev->logicalDpiY() / pm.logicalDpiY();
-
+    if (pdev && !pm.isNull())
+        scale = qreal(pdev->logicalDpiY()) / qreal(pm.logicalDpiY());
 
     QSize size = pm.size();
     if (hasWidth)
